djangoapp/requirements.txt
Django==5.2.8
psycopg2-binary==2.9.11
pillow>=12.0.0
django-summernote>=0.8.20.0
bleach>=6.3.0
django-axes>=8.0.0
python-dotenv>=1.2.1
